键鼠更新版（目前按住shift不能启动小陀螺，其他功能正常）

+++ Robot_Hero_bottom/User/Task/Chassis_task.h
@@ -18,9 +18,9 @@ typedef struct
 
 typedef struct
 {
-    pid_struct_t pid;  // 云台电机speed的pid结构体
-    fp32 pid_value[3]; // 云台电机speed的pid参数
-    fp32 target_speed; // 云台电机的目标速度
+    pid_struct_t pid;     // 云台电机speed的pid结构体
+    fp32 pid_value[3];    // 云台电机speed的pid参数
+    int16_t target_speed; // 云台电机的目标速度
 } chassis_t;
 
 // 获取imu——Yaw角度差值参数
